// lib/api/types/models/Response_GetDocumentResponse_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GetDocumentResponse } from './GetDocumentResponse';
export type Response_GetDocumentResponse_ = {
    status_code: number;
    description: string;
    data?: (GetDocumentResponse | null);
};